+++ src/components/Review.tsx
@@ -10,7 +10,8 @@ const Review: React.FC<Props> = ({ review }) => {
         <li key={review.name}>
             <figure className="relative flex flex-col rounded-lg p-6 shadow-lg border border-gray-50 bg-gradient bg-gradient-to-br from-gray-100 via-white to-gray-200">
                 <figcaption className="flex items-center space-x-4">
-                    <img src={review.img}
+                    <img
+                        src={review.img}
                         alt={review.companyName}
                         className="h-14 w-14 rounded-full shadow-md"
                     />
@@ -21,8 +22,8 @@ const Review: React.FC<Props> = ({ review }) => {
                     </div>
                 </figcaption>
 
-                <blockquote className="mt-5 text-gray-500 leading-6">
-                    <p>{review.review}</p>
+                <blockquote className="mt-5 text-gray-500 leading-6 [&>p]:mb-5 [&>p:last-child]:mb-0">
+                    {review.review}
                 </blockquote>
             </figure>
         </li>
